third_party/blink/web_tests/platform/linux-chrome/external/wpt/storage-access-api/storage-access-beyond-cookies.estimate.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for Quota
  assert_equals: Storage Access API should be accessible and return first-party data expected (string) "HasAccess for estimate" but got (object) object "[object Object]"
Harness: the test ran to completion.
